size_t

As per #70, we will run into problems with more than 2bn particles per node as all the particle counters and loop variables are currently `int`.

As a first step, switch the stored particle counts to `size_t`.

The next series of commits will be to change all the *loops* over the particles. I'm breaking this into two pieces since these changes are highly disruptive, especially since I'm submitting them along with !113.

See merge request !114

+++ src/debug.c
@@ -51,7 +51,7 @@
  */
 
 void printParticle(struct part *parts, struct xpart *xparts, long long int id,
-                   int N) {
+                   size_t N) {
 
   int i, found = 0;
 
@@ -66,7 +66,7 @@ void printParticle(struct part *parts, struct xpart *xparts, long long int id,
   if (!found) printf("## Particles[???] id=%lld not found\n", id);
 }
 
-void printgParticle(struct gpart *parts, long long int id, int N) {
+void printgParticle(struct gpart *parts, long long int id, size_t N) {
 
   int i, found = 0;
